change location of auth pages to change the route and updated imports
