+++ app/login/error/page.tsx
@@ -4,9 +4,9 @@ import siteInfo from "@/siteConfig";
 export default function Error() {
   return (
     <ErrorPage
-      title="Oops! Something went wrong"
-      description="Our busy bee couldn't create your account! Please try again."
-      redirectTo="/auth/signup"
+      title="Oops! Login Failed"
+      description="Our busy bee couldn't verify your credentials! Please try again."
+      redirectTo="/login"
       gif={siteInfo.errorGif}
     />
   );
