+++ app/login/page.tsx
@@ -1,4 +1,4 @@
-import SigninForm from "@/components/SigninForm";
+import SigninForm from "@/components/forms/SigninForm";
 
 export default function Login() {
   return (
